requirements.txt
streamlit
google-generativeai
pypdf
langchain
langchain-community
langchain-google-genai
chromadb